requirements.txt
pyzmq
click 

fastapi
uvicorn

tiktoken
protobuf
sentence-transformers